requirements.txt
tensorflow
pillow
numpy
tqdm
matplotlib
scikit-learn
